separated update script from database objects
